implement sharp attachments
