+++ src/talk/chat/attachment/chat-attachment.ts
@@ -342,361 +342,6 @@ export class LongTextAttachment implements ChatAttachment {
 
 }
 
-export enum SharpContentType {
-
-    NONE = '',
-    LIST = 'list',
-    IMAGE = 'image',
-    VIDEO_CLIP = 'vclip',
-    WEATHER = 'weather',
-    MOVIE = 'movie'
-
-}
-
-export abstract class SharpContent implements AttachmentContent {
-    
-    abstract readRawContent(rawData: any): void;
-    
-    abstract toRawContent(): any;
-
-}
-
-export abstract class SharpFragment extends SharpContent {
-
-}
-
-export class SharpImageFragment extends SharpFragment {
-
-    constructor(
-        public ImageURL: string = '',
-        public ImageWidth: number = -1,
-        public ImageHeight: number = -1
-    ) {
-        super();
-    }
-
-    readRawContent(rawData: any) {
-        this.ImageURL = rawData['I'] || '';
-        this.ImageWidth = rawData['W'] || -1;
-        this.ImageHeight = rawData['H'] || -1;
-    }
-
-    toRawContent(): any {
-        let obj: any = {
-            'I': this.ImageURL
-        }
-
-        if (this.ImageWidth !== -1) {
-            obj['W'] = this.ImageWidth;
-        }
-
-        if (this.ImageHeight !== -1) {
-            obj['H'] = this.ImageHeight;
-        }
-
-        return obj;
-    }
-}
-
-export class SharpTextFragment extends SharpFragment {
-    constructor(
-        public Text: string = '',
-        public Description: string = ''
-    ) {
-        super();
-    }
-
-    readRawContent(rawData: any) {
-        this.Text = rawData['T'] || '';
-        this.Description = rawData['DE'] || '';
-    }
-
-    toRawContent(): any {
-        let obj: any = {
-            'T': this.Text
-        }
-
-        if (this.Description !== '') {
-            obj['DE'] = this.Description;
-        }
-
-        return obj;
-    }
-}
-
-export class SharpAttachment implements ChatAttachment {
-
-    constructor(
-        public Question: string = '',
-        public RedirectURL: string = '',
-        public ContentType: SharpContentType = SharpContentType.NONE,
-        public ContentList: SharpContent[] = [],
-        public MainImage?: SharpImageFragment
-    ) {
-        
-    }
-
-    get RequiredMessageType() {
-        return MessageType.Search;
-    }
-
-    readAttachment(rawJson: any): void {
-        this.Question = rawJson['Q'];
-
-        this.ContentType = rawJson['V'] || SharpContentType.NONE;
-
-        if (rawJson['I']) {
-            this.MainImage = new SharpImageFragment();
-            this.MainImage.readRawContent(rawJson);
-        }
-
-        this.RedirectURL = rawJson['L'];
-
-        this.ContentList = [];
-
-        if (rawJson['R']) {
-            let list: any[] = rawJson['R'];
-
-            for (let rawContent of list) {
-                let content: SharpContent;
-
-                switch(this.ContentType) {
-                    case SharpContentType.VIDEO_CLIP: content = new SharpVideoContent(); break;
-                    
-                    case SharpContentType.LIST:
-                    case SharpContentType.IMAGE: content = new SharpImageContent(); break;
-
-                    case SharpContentType.MOVIE: content = new SharpMovieContent(); break;
-
-                    case SharpContentType.WEATHER: 
-                    default: continue;
-                }
-
-                content.readRawContent(rawContent);
-
-                this.ContentList.push(content);
-            }
-        }
-    }
-
-    toJsonAttachment() {
-        let obj: any = {
-            'Q': this.Question,
-            'V': this.ContentType,
-            'L': this.RedirectURL
-        };
-
-        if (this.MainImage) {
-            Object.assign(obj, this.MainImage.toRawContent());
-        }
-
-        if (this.ContentList.length > 0) {
-            let rawList = [];
-
-            for (let content of this.ContentList) {
-                rawList.push(content.toRawContent());
-            }
-
-            obj['R'] = rawList;
-        }
-
-        return obj;
-    }
-
-}
-
-export class SharpImageContent extends SharpContent {
-
-    constructor(
-        public RedirectURL: string = '',
-        public Text?: SharpTextFragment,
-        public InfoText: string = '',
-        public Image?: SharpImageFragment
-    ) {
-        super();
-    }
-
-    readRawContent(rawData: any) {
-        if (rawData['T']) {
-            this.Text = new SharpTextFragment();
-            this.Text.readRawContent(rawData);
-        }
-
-        this.InfoText = rawData['D'];
-
-        this.RedirectURL = rawData['L'];
-
-        if (rawData['I']) {
-            this.Image = new SharpImageFragment();
-            this.Image.readRawContent(rawData);
-        }
-        
-    }
-
-    toRawContent(): any {
-        let obj: any = {
-            'L': this.RedirectURL
-        };
-
-        if (this.InfoText !== '') {
-            obj['D'] = this.InfoText;
-        }
-
-        if (this.Text) {
-            Object.assign(obj, this.Text.toRawContent());
-        }
-
-        if (this.Image) {
-            Object.assign(obj, this.Image.toRawContent());
-        }
-
-        return obj;
-    }
-
-}
-
-export class SharpMovieContent extends SharpContent {
-
-    constructor(
-        public RedirectURL: string = '',
-        public Text?: SharpTextFragment,
-        public InfoText: string = '',
-        public StarRate: string = '',
-        public ExtraInfoList: SharpTextFragment[] = [],
-        public ImageList: SharpImageFragment[] = []
-    ) {
-        super();
-    }
-
-    readRawContent(rawData: any) {
-        if (rawData['T']) {
-            this.Text = new SharpTextFragment();
-            this.Text.readRawContent(rawData);
-        }
-
-        this.InfoText = rawData['D'];
-
-        this.RedirectURL = rawData['L'];
-
-        if (rawData['IL']) {
-            this.ImageList = [];
-
-            for (let rawImage of rawData['IL']) {
-                if (!rawImage) continue;
-                
-                let img = new SharpImageFragment();
-                img.readRawContent(rawImage);
-                this.ImageList.push(img);
-            }
-            
-        }
-
-        if (rawData['DL']) {
-            this.ExtraInfoList = [];
-
-            for (let rawText of rawData['DL']) {
-                if (!rawText) continue;
-                
-                let text = new SharpTextFragment();
-                text.readRawContent(rawText);
-                this.ExtraInfoList.push(text);
-            }
-        }
-        
-        this.StarRate = rawData['ST'];
-    }
-
-    toRawContent(): any {
-        let obj: any = {
-            'L': this.RedirectURL
-        };
-
-        if (this.InfoText !== '') {
-            obj['D'] = this.InfoText;
-        }
-
-        if (this.Text) {
-            Object.assign(obj, this.Text.toRawContent());
-        }
-
-        if (this.ImageList.length > 0) {
-            let list = [];
-            for (let image of this.ImageList) {
-                list.push(image.toRawContent());
-            }
-
-            obj['IL'] = list;
-        }
-
-        if (this.ExtraInfoList.length > 0) {
-            let list = [];
-            for (let text of this.ExtraInfoList) {
-                list.push(text.toRawContent());
-            }
-
-            obj['DL'] = list;
-        }
-
-        obj['ST'] = this.StarRate;
-
-        return obj;
-    }
-
-}
-
-export class SharpVideoContent extends SharpContent {
-
-    constructor(
-        public RedirectURL: string = '',
-        public Text?: SharpTextFragment,
-        public InfoText: string = '',
-        public PlayTime: number = 0,
-        public Image?: SharpImageFragment
-    ) {
-        super();
-    }
-
-    readRawContent(rawData: any) {
-        if (rawData['T']) {
-            this.Text = new SharpTextFragment();
-            this.Text.readRawContent(rawData);
-        }
-
-        this.InfoText = rawData['D'];
-
-        this.PlayTime = rawData['PT'] || 0;
-
-        if (rawData['I']) {
-            this.Image = new SharpImageFragment();
-            this.Image.readRawContent(rawData);
-        }
-
-        this.RedirectURL = rawData['L'];
-    }
-
-    toRawContent(): any {
-        let obj: any = {
-            'L': this.RedirectURL,
-            'PT': this.PlayTime
-        };
-
-        if (this.InfoText !== '') {
-            obj['D'] = this.InfoText;
-        }
-
-        if (this.Text) {
-            Object.assign(obj, this.Text.toRawContent());
-        }
-
-        if (this.Image) {
-            Object.assign(obj, this.Image.toRawContent());
-        }
-
-        return obj;
-    }
-
-}
-
 export class MentionContentList implements AttachmentContent {
     
     constructor(
